requirements.txt
pyyaml>=6.0
pydantic>=2.0
numpy>=1.24
pandas>=2.0

faiss-cpu>=1.7
qdrant-client>=1.7

sentence-transformers>=2.2
flagembedding>=1.2

python-dotenv>=1.0
httpx>=0.25

presidio-analyzer>=2.2
presidio-anonymizer>=2.2
presidio-pii-nlp-manager>=0.0

mlflow>=2.12
wandb>=0.16

feast>=0.40
redis>=5.0

deepEval>=0.4

pyarrow>=14.0
fastparquet>=2024.1

great-expectations>=0.18

pytest>=7.4
pytest-cov>=4.1

pre-commit>=3.6

dvc>=3.30
s3fs>=2023.12

airflow>=2.8
 astronomer-providers>=3.4

prometheus-client>=0.19
